Completed issue #5 - Successfully seperated the methods & classes code from the Form1.cs file.
